api/requirements.txt
langchain
python-dotenv
openai
chromadb
flask